--- examples/workflows/class-based/image_transforms/ImageTransforms.json
@@ -1,1 +1,0 @@
-{"version": "0", "type": "workflow", "nodes": [{"id": "4", "type": "step", "position": {"x": 200, "y": 200}, "data": {"name": "HuggingfaceDataset", "parameters": {"dataset_id": {"type": "string", "description": "The dataset ID from Huggingface", "value": "kmeng/daisy-dog"}, "split": {"type": "string", "default": "train", "description": "The split of the dataset to use", "value": "train"}, "shuffle": {"type": "boolean", "default": false, "description": "Whether to shuffle the dataset", "value": false}, "log_data": {"type": "boolean", "default": true, "description": "Whether to log the outputs as JSON to the node UI", "value": true}, "image_columns": {"type": "list[string]", "description": "The columns in the dataset that contain images. This is to let Graphbook know how to display the images in the UI.", "required": false}, "kwargs": {"type": "dict", "description": "Additional keyword arguments to pass to the dataset", "required": false, "value": []}}, "inputs": [], "outputs": ["out"], "category": "Huggingface", "module": "graphbook_huggingface.hf_datasets", "label": "HuggingfaceDataset", "isCollapsed": false}}, {"id": "5", "type": "step", "position": {"x": 407, "y": 200}, "data": {"name": "HorizontalFlip", "parameters": {"p": {"type": "number", "default": 0.5, "value": 0.5}}, "inputs": ["in"], "outputs": ["out"], "category": "Transform", "module": "transient_module", "label": "HorizontalFlip", "isCollapsed": false}}, {"id": "7", "type": "step", "position": {"x": 820, "y": 200}, "data": {"name": "ColorJitter", "parameters": {"brightness": {"type": "number", "default": 0.5, "value": 0.5}, "contrast": {"type": "number", "default": 0.5, "value": 0.5}, "saturation": {"type": "number", "default": 0.5, "value": 0.5}, "hue": {"type": "number", "default": 0.5, "value": 0.5}}, "inputs": ["in"], "outputs": ["out"], "category": "Transform", "module": "transient_module", "label": "ColorJitter", "isCollapsed": false}}, {"id": "8", "type": "step", "position": {"x": 614, "y": 200}, "data": {"name": "ResizedCrop", "parameters": {"size": {"type": "array", "default": [224, 224], "value": [224, 224]}, "scale": {"type": "array", "default": [0.08, 1], "value": [0.08, 1]}, "ratio": {"type": "array", "default": [0.75, 1.3333333333333333], "value": [0.75, 1.3333333333333333]}}, "inputs": ["in"], "outputs": ["out"], "category": "Transform", "module": "transient_module", "label": "ResizedCrop", "isCollapsed": false}}, {"id": "9", "type": "step", "position": {"x": 1029, "y": 192}, "data": {"name": "Normalize", "parameters": {"mean": {"type": "array", "default": [0.485, 0.456, 0.406], "value": [0.485, 0.456, 0.406]}, "std": {"type": "array", "default": [0.229, 0.224, 0.225], "value": [0.229, 0.224, 0.225]}}, "inputs": ["in"], "outputs": ["out"], "category": "Transform", "module": "transient_module", "label": "Normalize", "isCollapsed": false}}], "edges": [{"source": "4", "sourceHandle": "out", "target": "5", "targetHandle": "in", "data": {}, "id": "reactflow__edge-4out-5in"}, {"source": "5", "sourceHandle": "out", "target": "8", "targetHandle": "in", "data": {}, "id": "reactflow__edge-5out-8in"}, {"source": "8", "sourceHandle": "out", "target": "7", "targetHandle": "in", "data": {}, "id": "reactflow__edge-8out-7in"}, {"source": "7", "sourceHandle": "out", "target": "9", "targetHandle": "in", "data": {}, "id": "reactflow__edge-7out-9in"}]}
