--- examples/workflows/class-based/image_transforms/docs/ImageTransforms.md
@@ -1,5 +1,0 @@
-# Image Transforms
-
-This workflow completes a series of image transformations on incoming images. It
-leverages the [graphbook_huggingface](https://github.com/graphbookai/graphbook-huggingface) plugin
-for downloading sample image data.
