+++ examples/workflows/class-based/web_scraper/PageScraper.json
@@ -1,1 +1,1 @@
-{"version": "0", "type": "workflow", "nodes": [{"id": "0", "type": "step", "position": {"x": 242, "y": 196}, "data": {"name": "UrlSource", "parameters": {"url": {"type": "string", "default": "https://www.graphbook.ai/", "value": "https://www.graphbook.ai/"}}, "inputs": [], "outputs": ["url"], "category": "Custom", "module": "transient_module", "label": "UrlSource", "isCollapsed": false}}, {"id": "1", "type": "resource", "position": {"x": 247, "y": 308}, "data": {"name": "PersonalInfo", "parameters": {"api_key": {"type": "string", "default": "<API_KEY_HERE>", "value": "<API_KEY_HERE>"}}, "category": "Custom", "module": "transient_module", "label": "PersonalInfo", "isCollapsed": false}}, {"id": "2", "type": "step", "position": {"x": 479, "y": 244}, "data": {"name": "ScrapePage", "parameters": {"api_key": {"type": "resource"}}, "inputs": ["in"], "outputs": ["mission_statement"], "category": "Custom", "module": "transient_module", "label": "ScrapePage", "isCollapsed": false}}, {"id": "3", "type": "export", "position": {"x": 683, "y": 277}, "data": {"label": "Output", "exportType": "output", "isResource": false, "isCollapsed": false}}], "edges": [{"source": "0", "sourceHandle": "url", "target": "2", "targetHandle": "in", "data": {}, "id": "reactflow__edge-0url-2in"}, {"source": "1", "sourceHandle": "resource", "target": "2", "targetHandle": "api_key", "data": {}, "id": "reactflow__edge-1resource-2api_key"}, {"source": "2", "sourceHandle": "mission_statement", "target": "3", "targetHandle": "out", "data": {}, "id": "reactflow__edge-2mission_statement-3out"}]}
+{"version": "0", "type": "workflow", "nodes": [{"id": "0", "type": "step", "position": {"x": 242, "y": 196}, "data": {"name": "UrlSource", "parameters": {"url": {"type": "string", "default": "https://www.graphbook.ai/", "value": "https://www.graphbook.ai/"}}, "inputs": [], "outputs": ["url"], "category": "Custom", "module": "transient_module", "label": "UrlSource", "isCollapsed": false}}, {"id": "1", "type": "resource", "position": {"x": 247, "y": 308}, "data": {"name": "PersonalInfo", "parameters": {"api_key": {"type": "string", "default": "<API_KEY_HERE>", "value": "<API_KEY_HERE>"}}, "category": "Custom", "module": "transient_module", "label": "PersonalInfo", "isCollapsed": false}}, {"id": "3", "type": "export", "position": {"x": 683, "y": 277}, "data": {"label": "Output", "exportType": "output", "isResource": false, "isCollapsed": false}}, {"id": "4", "type": "step", "position": {"x": 463, "y": 241}, "data": {"name": "ScrapeMission", "parameters": {"api_key": {"type": "resource"}}, "inputs": ["in"], "outputs": ["out"], "category": "Custom", "module": "transient_module", "label": "ScrapeMission", "isCollapsed": false}}], "edges": [{"source": "0", "sourceHandle": "url", "target": "4", "targetHandle": "in", "data": {}, "id": "reactflow__edge-0url-4in"}, {"source": "1", "sourceHandle": "resource", "target": "4", "targetHandle": "api_key", "data": {}, "id": "reactflow__edge-1resource-4api_key"}, {"source": "4", "sourceHandle": "out", "target": "3", "targetHandle": "out", "data": {}, "id": "reactflow__edge-4out-3out"}]}
