+++ examples/workflows/class-based/web_scraper/WebsiteScraper.json
@@ -1,1 +1,1 @@
-{"version": "0", "type": "workflow", "nodes": [{"id": "0", "type": "step", "position": {"x": 225, "y": 215}, "data": {"name": "UrlSource", "parameters": {"url": {"type": "string", "default": "https://www.graphbook.ai/", "value": "https://www.graphbook.ai/"}}, "inputs": [], "outputs": ["url"], "category": "Custom", "module": "transient_module", "label": "UrlSource", "isCollapsed": false}}, {"id": "1", "type": "step", "position": {"x": 452, "y": 299}, "data": {"name": "ScrapeWebsite", "parameters": {"api_key": {"type": "resource"}, "max_depth": {"type": "resource"}}, "inputs": ["in"], "outputs": ["page_titles"], "category": "Custom", "module": "transient_module", "label": "ScrapeWebsite", "isCollapsed": false}}, {"id": "2", "type": "resource", "position": {"x": 229, "y": 444}, "data": {"name": "CrawlParams", "parameters": {"max_depth": {"type": "number", "default": 3, "value": 3}}, "category": "Custom", "module": "transient_module", "label": "CrawlParams", "isCollapsed": false}}, {"id": "3", "type": "resource", "position": {"x": 240, "y": 333}, "data": {"name": "PersonalInfo", "parameters": {"api_key": {"type": "string", "default": "<API_KEY_HERE>", "value": "<API_KEY_HERE>"}}, "category": "Custom", "module": "transient_module", "label": "PersonalInfo", "isCollapsed": false}}, {"id": "4", "type": "export", "position": {"x": 695, "y": 332}, "data": {"label": "Output", "exportType": "output", "isResource": false, "isCollapsed": false}}], "edges": [{"source": "0", "sourceHandle": "url", "target": "1", "targetHandle": "in", "data": {}, "id": "reactflow__edge-0url-1in"}, {"source": "3", "sourceHandle": "resource", "target": "1", "targetHandle": "api_key", "data": {}, "id": "reactflow__edge-3resource-1api_key"}, {"source": "2", "sourceHandle": "resource", "target": "1", "targetHandle": "max_depth", "data": {}, "id": "reactflow__edge-2resource-1max_depth"}, {"source": "1", "sourceHandle": "page_titles", "target": "4", "targetHandle": "out", "data": {}, "id": "reactflow__edge-1page_titles-4out"}]}
+{"version": "0", "type": "workflow", "nodes": [{"id": "3", "type": "resource", "position": {"x": 232.86542881634045, "y": 352.4261420815885}, "data": {"name": "PersonalInfo", "parameters": {"api_key": {"type": "string", "default": "<API_KEY_HERE>", "value": "<API_KEY_HERE>"}}, "category": "Custom", "module": "transient_module", "label": "PersonalInfo", "isCollapsed": false}}, {"id": "4", "type": "export", "position": {"x": 695, "y": 332}, "data": {"label": "Output", "exportType": "output", "isResource": false, "isCollapsed": false}}, {"id": "7", "type": "step", "position": {"x": 225, "y": 226}, "data": {"name": "WebCrawlGenerator", "parameters": {"url": {"type": "string", "default": "https://www.graphbook.ai/", "value": "https://www.graphbook.ai/"}, "max_depth": {"type": "number", "default": 3, "value": 3}}, "inputs": [], "outputs": ["out"], "category": "Custom", "module": "transient_module", "label": "WebCrawlGenerator", "isCollapsed": false}}, {"id": "8", "type": "step", "position": {"x": 478, "y": 306}, "data": {"name": "ScrapeTitle", "parameters": {"api_key": {"type": "resource"}}, "inputs": ["in"], "outputs": ["out"], "category": "Custom", "module": "transient_module", "label": "ScrapeTitle", "isCollapsed": false}}], "edges": [{"source": "7", "sourceHandle": "out", "target": "8", "targetHandle": "in", "data": {}, "id": "reactflow__edge-7out-8in"}, {"source": "3", "sourceHandle": "resource", "target": "8", "targetHandle": "api_key", "data": {}, "id": "reactflow__edge-3resource-8api_key"}, {"source": "8", "sourceHandle": "out", "target": "4", "targetHandle": "out", "data": {}, "id": "reactflow__edge-8out-4out"}]}
