+++ examples/workflows/functional/web_scraper/PageScraper.json
@@ -1,1 +1,1 @@
-{"version": "0", "type": "workflow", "nodes": [{"id": "0", "type": "step", "position": {"x": 195, "y": 167}, "data": {"name": "UrlSource", "parameters": {"url": {"type": "string", "default": "https://www.graphbook.ai/", "required": false, "description": "", "value": "https://www.graphbook.ai/"}}, "inputs": [], "outputs": ["url"], "category": "Custom", "module": "graphbook.core.decorators", "label": "UrlSource", "isCollapsed": false}}, {"id": "1", "type": "step", "position": {"x": 444, "y": 266}, "data": {"name": "ScrapePage", "parameters": {"api_key": {"type": "resource", "default": null, "required": false, "description": "", "value": null}}, "inputs": ["in"], "outputs": ["mission_statement"], "category": "Custom", "module": "graphbook.core.decorators", "label": "ScrapePage", "isCollapsed": false}}, {"id": "2", "type": "resource", "position": {"x": 208, "y": 330}, "data": {"name": "PersonalInfo", "parameters": {"api_key": {"type": "string", "default": "<API_KEY_HERE>", "required": false, "description": "", "value": "<API_KEY_HERE>"}}, "category": "Custom", "module": "graphbook.core.decorators", "label": "PersonalInfo", "isCollapsed": false}}, {"id": "3", "type": "export", "position": {"x": 672, "y": 302}, "data": {"label": "Output", "exportType": "output", "isResource": false, "isCollapsed": false}}], "edges": [{"source": "0", "sourceHandle": "url", "target": "1", "targetHandle": "in", "data": {}, "id": "reactflow__edge-0url-1in"}, {"source": "2", "sourceHandle": "resource", "target": "1", "targetHandle": "api_key", "data": {}, "id": "reactflow__edge-2resource-1api_key"}, {"source": "1", "sourceHandle": "mission_statement", "target": "3", "targetHandle": "out", "data": {}, "id": "reactflow__edge-1mission_statement-3out"}]}
+{"version": "0", "type": "workflow", "nodes": [{"id": "0", "type": "step", "position": {"x": 195, "y": 167}, "data": {"name": "UrlSource", "parameters": {"url": {"type": "string", "default": "https://www.graphbook.ai/", "required": false, "description": "", "value": "https://www.graphbook.ai/"}}, "inputs": [], "outputs": ["url"], "category": "Custom", "module": "graphbook.core.decorators", "label": "UrlSource", "isCollapsed": false}}, {"id": "2", "type": "resource", "position": {"x": 208, "y": 330}, "data": {"name": "PersonalInfo", "parameters": {"api_key": {"type": "string", "default": "<API_KEY_HERE>", "required": false, "description": "", "value": "<API_KEY_HERE>"}}, "category": "Custom", "module": "graphbook.core.decorators", "label": "PersonalInfo", "isCollapsed": false}}, {"id": "3", "type": "export", "position": {"x": 672, "y": 287}, "data": {"label": "Output", "exportType": "output", "isResource": false, "isCollapsed": false}}, {"id": "4", "type": "step", "position": {"x": 407, "y": 262}, "data": {"name": "ScrapeMission", "parameters": {"api_key": {"type": "resource", "default": null, "required": false, "description": "", "value": null}}, "inputs": ["in"], "outputs": ["out"], "category": "Custom", "module": "graphbook.core.decorators", "label": "ScrapeMission", "isCollapsed": false}}], "edges": [{"source": "0", "sourceHandle": "url", "target": "4", "targetHandle": "in", "data": {}, "id": "reactflow__edge-0url-4in"}, {"source": "2", "sourceHandle": "resource", "target": "4", "targetHandle": "api_key", "data": {}, "id": "reactflow__edge-2resource-4api_key"}, {"source": "4", "sourceHandle": "out", "target": "3", "targetHandle": "out", "data": {}, "id": "reactflow__edge-4out-3out"}]}
