+++ examples/workflows/functional/web_scraper/WebsiteScraper.json
@@ -1,1 +1,1 @@
-{"version": "0", "type": "workflow", "nodes": [{"id": "0", "type": "step", "position": {"x": 199, "y": 188}, "data": {"name": "UrlSource", "parameters": {"url": {"type": "string", "default": "https://www.graphbook.ai/", "required": false, "description": "", "value": "https://www.graphbook.ai/"}}, "inputs": [], "outputs": ["url"], "category": "Custom", "module": "graphbook.core.decorators", "label": "UrlSource", "isCollapsed": false}}, {"id": "1", "type": "step", "position": {"x": 439, "y": 261}, "data": {"name": "ScrapeWebsite", "parameters": {"api_key": {"type": "resource", "default": null, "required": false, "description": "", "value": null}, "max_depth": {"type": "resource", "default": null, "required": false, "description": "", "value": null}}, "inputs": ["in"], "outputs": ["page_titles"], "category": "Custom", "module": "graphbook.core.decorators", "label": "ScrapeWebsite", "isCollapsed": false}}, {"id": "2", "type": "resource", "position": {"x": 206, "y": 281}, "data": {"name": "PersonalInfo", "parameters": {"api_key": {"type": "string", "default": "<API_KEY_HERE>", "required": false, "description": "", "value": "<API_KEY_HERE>"}}, "category": "Custom", "module": "graphbook.core.decorators", "label": "PersonalInfo", "isCollapsed": false}}, {"id": "4", "type": "export", "position": {"x": 704, "y": 300}, "data": {"label": "Output", "exportType": "output", "isResource": false, "isCollapsed": false}}, {"id": "5", "type": "resource", "position": {"x": 198, "y": 371}, "data": {"name": "CrawlParams", "parameters": {"max_depth": {"type": "number", "default": 3, "required": false, "description": "", "value": 3}}, "category": "Custom", "module": "graphbook.core.decorators", "label": "CrawlParams", "isCollapsed": false}}], "edges": [{"source": "2", "sourceHandle": "resource", "target": "1", "targetHandle": "api_key", "data": {}, "id": "reactflow__edge-2resource-1api_key"}, {"source": "0", "sourceHandle": "url", "target": "1", "targetHandle": "in", "data": {}, "id": "reactflow__edge-0url-1in"}, {"source": "1", "sourceHandle": "page_titles", "target": "4", "targetHandle": "out", "data": {}, "id": "reactflow__edge-1page_titles-4out"}, {"source": "5", "sourceHandle": "resource", "target": "1", "targetHandle": "max_depth", "data": {}, "id": "reactflow__edge-5resource-1max_depth"}]}
+{"version": "0", "type": "workflow", "nodes": [{"id": "2", "type": "resource", "position": {"x": 206, "y": 325}, "data": {"name": "PersonalInfo", "parameters": {"api_key": {"type": "string", "default": "<API_KEY_HERE>", "required": false, "description": "", "value": "<API_KEY_HERE>"}}, "category": "Custom", "module": "graphbook.core.decorators", "label": "PersonalInfo", "isCollapsed": false}}, {"id": "4", "type": "export", "position": {"x": 704, "y": 300}, "data": {"label": "Output", "exportType": "output", "isResource": false, "isCollapsed": false}}, {"id": "6", "type": "step", "position": {"x": 194, "y": 164}, "data": {"name": "WebScrawlGenerator", "parameters": {"url": {"type": "string", "default": "https://www.graphbook.ai/", "required": false, "description": "", "value": "https://www.graphbook.ai/"}, "max_depth": {"type": "number", "default": 3, "required": false, "description": "", "value": 3}}, "inputs": [], "outputs": ["out"], "category": "Custom", "module": "graphbook.core.decorators", "label": "WebScrawlGenerator", "isCollapsed": false}}, {"id": "7", "type": "step", "position": {"x": 454, "y": 276}, "data": {"name": "ScrapeTitle", "parameters": {"api_key": {"type": "resource", "default": null, "required": false, "description": "", "value": null}}, "inputs": ["in"], "outputs": ["out"], "category": "Custom", "module": "graphbook.core.decorators", "label": "ScrapeTitle", "isCollapsed": false}}], "edges": [{"source": "6", "sourceHandle": "out", "target": "7", "targetHandle": "in", "data": {}, "id": "reactflow__edge-6out-7in"}, {"source": "2", "sourceHandle": "resource", "target": "7", "targetHandle": "api_key", "data": {}, "id": "reactflow__edge-2resource-7api_key"}, {"source": "7", "sourceHandle": "out", "target": "4", "targetHandle": "out", "data": {}, "id": "reactflow__edge-7out-4out"}]}
